Initial commit
Create a new project, "Pulp Fixtures." This new project is a collection
of scripts for creating fixture data for a Pulp server. At present,
Docker and RPM fixture data can be generated with the following
commands, respectively:

    make docker-fixtures
    make rpm-fixtures

Each of these make targets requires that certain utilities be available
and usable. For example, `docker-fixtures` requires that the Docker
tools are installed, that the Docker daemon is running and that the
current user belongs to the `docker` group.

The advantage of collecting these scripts into a single repository is
that the scripts can be given a fairly common interface and set of
behaviours. In turn, this allows users of tools like Pulp Smash to more
easily reproduce a test environment.

The Docker scripts are written in response to
https://github.com/PulpQE/pulp-smash/issues/81

The RPM scripts are adapted from
https://github.com/PulpQE/pulp-smash/pull/221 (Thanks @bmbouter!)