Write all fixtures to a single directory
Add a top-level directory named `fixtures`, and update the make file so
that all compiled files are placed in this directory. **This change is
breaking.** If a client used this command before this commit:

    make docker-fixtures

Then the client should use this command after this commit:

    make fixtures/docker

The former command created a directory named `docker-fixtures`, and the
latter command creates a directory at `fixtures/docker`.

The `all` target has been kept for backward compatibility, but it issues
a warning when successfully executed.

Fix https://github.com/PulpQE/pulp-fixtures/issues/6