Make some nitpicky changes to the new make targets
The createrepo and createrepo_c man pages list a `--deltas` option, but
not a `--delta` option. Use the one listed in the man page. This change
has no visible effect, indicating that `--delta` is supported for legacy
reasons.

Re-order the make targets, make target implementations, and readme
descriptions of the make targets, so that all are consistent and in
alphabetical order.

Update some prose.