Fix bug in RPM package signing procedure
The `fixtures/rpm`, `fixtures/drpm` and `fixtures/srpm` make targets
generate repositories with incorrect metadata. The issue is that, when
creating a signed repository, packages and metadata are written to the
destination directory before package signing is performed. When packages
are signed, the existing metadata is no longer accurate. Of special
interest is that package checksums and sizes are no longer accurate.

Fix the `fixtures/rpm` and `fixtures/srpm` make targets, so that package
signing is performed before repository metadata is generated.

See: https://github.com/PulpQE/pulp-fixtures/issues/32