De-duplicate and normalize comments
Both the readme and the makefile contain comments on how to customize
the base URL of the mirrorlist targets and the syntax of the mirrorlist
targets. De-duplicate this information:

* Let information on how to customize the base URL of mirrorlist targets
  live in the makefile. This is appropriate because other make target
  usage information also lives here.
* Let notes regarding the syntax of mirrorlist files live in the readme
  file. This is appropriate because other caveats are also listed as
  notes in the readme.

Finally, make all of the usage caveats in the readme look similar.

It may be a good idea to create a section dedicated to usage caveats.