Rename target "python" to "python-pulp"
Pulp is currently able to sync Python packages from PyPI. As of the
pulp_python plugin version 2.0, it should also be able to sync Python
packages from other Pulp systems. As a result, Pulp Fixtures should host
both types of repositories for testing purposes.

Rename the "python" target to "python-pulp." This new name more clearly
indicates the purpose of this target: to generate a repository like what
Pulp generates. In the future, a target named "python-pypi" may be
added, and having a pair of targets named "python-pulp" and
"python-pypi" is nicely unambiguous.

Keep the "python" target. Remove all documented references to this
target, in an attempt to deprecate it. Let it be a symlink to the
repository generated by the "python-pulp" target.