Add file fixtures
File fixture is a repository where Pulp will publish the files. On Pulp
2 it is imported by the iso command but it does not mean that the files
should be a real ISO, it can be any file type. On Pulp 3 it will be
renamed to file and that is why this is called file fixtures.

A valid file fixture repository must have the fixtures files and the
PULP_MANIFEST file, the latter is a CSV file where each line is the file
path relative to the fixtures URL, the sha256sum of the file and the
size, in bytes, of the file.

Also add the file-mixed fixture which will generate a file fixture with
valid and invalid entries on the PULP_MANIFEST file.