Target a consistent platform when generating repos
`fedpkg local` lets one create a "local test rpmbuild binary" from a
spec file. Several fixtures targets use this command, including
`fixtures/rpm-with-non-ascii`.

The `--release` lets one target a specific distribution, such as "f25."
Currently, this flag is passed an option depending on the distribution
on which packages are being generated. For example, if fixtures are
being generated on Fedora 25, then "f25" is passed, and if fixtures are
being generated on Fedora 26, then "f26" is passed. This behaviour seems
like a recipe for breakage: if we ever change what kind of VMs are used
to generate fixtures, then different RPMs will be generated, and clients
that depend on the fixtures may break.

Hard-code a target platform into the RPM generation script. This allows
the fixtures to be generated on a variety of distributions, while
producing the same result. The downside of this change is that the RPM
fixtures repository will require some maintenance over time, as the
target platform will eventually become extremely out of date. But the
benefit seems to greatly outweigh the downside.