Create new fixtures for python-pypi
These new fixtures are not hardcoded, instead we read in the new
projects.json file that includes a list of python projects, and the
distributions to keep for those projects.

The metadata is obtained from http://pypi.org/pypi/{project}/json
The packages are obtained from the metadata distribution url.

closes #3608
https://pulp.plan.io/issues/3608