Upgrade from Ubuntu Trusty to Xenial
Ubuntu Trusty is now four years old. There's little apparent reason to
stay with a four-year-old release of Ubuntu, and upgrading grants access
to newer packages. Of special interest is a newer version of Python:
Python scripts were recently added, and if any linting is to be done, it
should be done with a more current interpreter.