Adding srpm-duplicates fixture to address #4459
Issue #4459 addresses when there is duplicate SRPMs in the repodata
that pulp would error out in the sync when `--skip srpm` was used in
repo creation.

The following are the modifications:
- Adding two new SRPMS from .spec that are used for testing
- Updating Makefile to create two separate folders containing RPMs
- Run createrepo_c on the make to have repodata have two entries
  for the same RPM version in the primary.yaml
- Updated README with information on the new fixture

Note that this test fixture will have no SRPMS or RPMs copied as
part of the test. Only error messages and lack of content units
will be verified.

refs #4459