Commenting out the creation of pulp-fixtures puppet
There is an issue generating the puppet fixtures that is not
a trivial fix. The work to add puppet fixture generation is being
tracked by SATQE-3469.

There is an issue generating an RPM fixture with UTF-8 that is not
a trivial fix. The Work to add back this one generation is being
tracked by SATQE-3484.

This commit gets the building of all other fixtures functioning
and rsyncing again with a newer (F29) base OS. This work paves
the way to update soon to F30 which unblocks
an outstanding PR/functional test for createrepo_c testing in:
- https://pulp.plan.io/issues/4530
- https://github.com/PulpQE/pulp-fixtures/issues/120

See: https://projects.engineering.redhat.com/browse/SATQE-3467