Split function of the entry point in 2
This permit to build the new container in openshift,
and then run it in 2 steps.

When building the image in Openshift, the BASE_URL variable
can be defined at build time (in the BuildConfig object), and
so will run sed, but not nginx.

Then at deployment time, the URL having been set already, we do not
change anything.

[noissue]